requirements.txt
discord.py=1.7.3
python-dotenv=0.18.0
motor=2.4.0
dnspython=2.1.0